refactor Hr component for use in Our work page & move into common component folder
